minor fixes

+++ include/pose_calculations.hpp
@@ -32,7 +32,7 @@ enum ArenaDirection{RIGHT=0, FAR_SIDE, LEFT, NEAR_SIDE, TRUE_NORTH};
 
 pose_t transform_pose(const CoordinateSystem co,
                       const ArenaDirection co_north,
-                      const double true_north_deg,
+                      const float true_north_deg,
                       const UpAxis up_axis,
                       const ArenaDirection long_edge,
                       const ArenaDirection craft_nose,
